Make timezone removal migration safe if column absent

+++ db/migrate/20260719172035_remove_timezone_from_users.rb
@@ -1,5 +1,7 @@
 class RemoveTimezoneFromUsers < ActiveRecord::Migration[8.1]
   def change
-    remove_column :users, :timezone, :string
+    if column_exists?(:users, :timezone)
+      remove_column :users, :timezone, :string
+    end
   end
 end
